+++ src/test/calculadora/SumaTest.java
@@ -1,0 +1,13 @@
+package calculadora;
+
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+public class SumaTest {
+    private final Calculadora calc = new Calculadora();
+
+    @Test
+    void suma(){
+        assertEquals(calc.suma(5,3), 8.0);
+    }
+}
